+++ src/prop-keys.ts
@@ -1,0 +1,9 @@
+
+// use symbols to avoid name collisions in prototype chain
+export const scopeKey = Symbol('scope')
+export const modelKey = Symbol('model')
+export const actionsKey = Symbol('actions')
+export const actionIdsKey = Symbol('actionIds')
+export const actionStatesKey = Symbol('actionStates')
+export const actionsSizeKey = Symbol('actionsSize')
+export const watchStopHandlersKey = Symbol('watchStopHandlers')